Move verifyExtensionBundle
